+++ src/us/kbase/workspace/kbase/HandleIdHandlerFactory.java
@@ -31,20 +31,23 @@ public class HandleIdHandlerFactory implements IdReferenceHandlerFactory {
 	public static final IdReferenceType type = new IdReferenceType("handle");
 	private final URL handleService;
 	private final AuthToken userToken;
+	private final URL authURL;
 	
 	/** pass in null for the handle service URL to cause an exception to be
 	 * thrown if a handle id is encountered
 	 */
 	public HandleIdHandlerFactory(
 			final URL handleServiceURL,
-			final AuthToken userToken) {
+			final AuthToken userToken,
+			final URL authServiceURL) {
 		
-		if (userToken == null) {
+		if (userToken == null || authServiceURL == null) {
 			throw new NullPointerException(
-					"userToken cannot be null");
+					"userToken and authServiceURL cannot be null");
 		}
 		this.handleService = handleServiceURL;
 		this.userToken = userToken;
+		this.authURL = authServiceURL;
 	}
 	
 	@Override
@@ -115,7 +118,7 @@ public class HandleIdHandlerFactory implements IdReferenceHandlerFactory {
 			final Long allreadable;
 			try {
 				final AbstractHandleClient ahc = new AbstractHandleClient(
-						handleService, userToken);
+						handleService, userToken, authURL);
 				if (handleService.getProtocol().equals("http")) {
 					ahc.setIsInsecureHttpConnectionAllowed(true);
 				}