+++ src/us/kbase/workspace/test/database/mongo/GridFSBlobStoreTest.java
@@ -59,7 +59,7 @@ public class GridFSBlobStoreTest {
 	@AfterClass
 	public static void tearDownClass() throws Exception {
 		if (mongo != null) {
-			mongo.destroy(TestCommon.deleteTempFiles());
+			mongo.destroy(TestCommon.getDeleteTempFiles());
 		}
 	}